TicketBack/requirements.txt
APScheduler==3.10.0
asgi-redis==1.1.0
asgiref==3.6.0
astroid==2.12.13
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.12.1
Automat==22.10.0
boto3==1.26.74
botocore==1.29.74
certifi==2022.12.7
cffi==1.15.1
channels==3.0.5
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.1.0
cryptography==38.0.4
daphne==3.0.1
dill==0.3.6
dj-rest-auth==3.0.0
Django==4.1.4
django-allauth==0.52.0
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-storages==1.13.2
django-filter==22.1
django-health-check==3.17.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-model-utils==4.3.1
django-paypal==2.0
djangochannelsrestframework==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
exceptiongroup==1.1.0
facebook-sdk==3.1.0
google-api-python-client==2.84.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
gunicorn==20.0.4
h11==0.14.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
lazy-object-proxy==1.8.0
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.7.0
mercadopago==2.2.0
msgpack==1.0.4
msgpack-python==0.5.6
openapi-codec==1.3.2
outcome==1.2.0
packaging==22.0
pdfkit==1.0.0
Pillow==9.3.0
pkgutil-resolve-name==1.3.10
platformdirs==2.5.4
pluggy==1.0.0
psycopg2-binary==2.9.2
py3dns==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pylint==2.15.8
pyOpenSSL==22.1.0
pyrsistent==0.19.2
PySocks==1.7.1
pytest==7.2.1
pytest-asyncio==0.20.3
pytest-cov==4.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
redis==4.4.2
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
segno==1.5.2
selenium==4.8.0
service-identity==21.1.0
simplejson==3.18.0
singlemodeladmin==0.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
supervisor==4.2.2
tomli==2.0.1
tomlkit==0.11.6
trio==0.22.0
trio-websocket==0.9.2
txaio==22.2.1
typing-extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.13
uvicorn==0.20.0
validate-email==1.3
websocket-client==1.4.2
wrapt==1.14.1
wsproto==1.2.0
XlsxWriter==3.1.0
yapf==0.32.0
zipp==3.11.0
zope.interface==5.5.2